requirements.txt
pillow
bleak
pyserial